services/single-llm-api/requirements.txt
accelerate==0.31.0
fastapi==0.111.0
gunicorn==22.0.0
langchain-huggingface==0.0.3
pytest-cov==5.0.0
transformers==4.41.2
pyyaml==6.0.1
